fix(teleport): 修复实验延时换算编译错误

+++ BetterGenshinImpact/GameTask/AutoTrackPath/ExperimentalTeleportUiStateMachine.cs
@@ -443,7 +443,7 @@ internal sealed class ExperimentalTeleportUiStateMachine
             1,
             (int)Math.Round(
                 baseDelay * configured /
-                TpConfig.DefaultTeleportOperationDelayMilliseconds));
+                (double)TpConfig.DefaultTeleportOperationDelayMilliseconds));
     }
 
     private int GetStateRecognitionInterval() => Math.Clamp(
